react-research-agent/requirements.txt
beautifulsoup4>=4.12.0
langchain>=0.3.0
langchain-openai>=0.2.0
langsmith>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
tavily-python>=0.4.0
wikipedia>=1.4.0
pytest>=8.0.0
